cleaned a few things up tring to investigate crash (populse/mia_processes#11)
it crashes at the end of a workflow execution whenever an output value
(float for instance) is input into a later process.
It still crashes. I could not find out. The crash happens when we
remove/hide/do somethong to the RunProgress object, which is apparently
unrelated, thus suggest a thread problem somewhere.